+++ tests/posix/timers/src/clock.c
@@ -4,6 +4,10 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
+
+/* for tp_ge(), tp_diff() */
+#include "posix_clock.h"
+
 #include <sys/time.h>
 #include <time.h>
 #include <unistd.h>
@@ -28,35 +32,6 @@ static const bool settable[] = {
 	false,
 	true,
 };
-
-static inline int64_t ts_to_ns(const struct timespec *ts)
-{
-	return ts->tv_sec * NSEC_PER_SEC + ts->tv_nsec;
-}
-
-#define _tp_op(_a, _b, _op) (ts_to_ns(_a) _op ts_to_ns(_b))
-
-#define _decl_op(_type, _name, _op)                                                                \
-	__used static inline _type _name(const struct timespec *_a, const struct timespec *_b)     \
-	{                                                                                          \
-		return _tp_op(_a, _b, _op);                                                        \
-	}
-
-_decl_op(bool, tp_eq, ==);     /* a == b */
-_decl_op(bool, tp_lt, <);      /* a < b */
-_decl_op(bool, tp_gt, >);      /* a > b */
-_decl_op(bool, tp_le, <=);     /* a <= b */
-_decl_op(bool, tp_ge, >=);     /* a >= b */
-_decl_op(int64_t, tp_diff, -); /* a - b */
-
-/* lo <= (a - b) < hi */
-__used static inline bool tp_diff_in_range_ns(const struct timespec *a, const struct timespec *b,
-					      int64_t lo, int64_t hi)
-{
-	int64_t diff = tp_diff(a, b);
-
-	return diff >= lo && diff < hi;
-}
 
 ZTEST(posix_timers, test_clock_gettime)
 {
